Validation of backend specific configuration
