+++ logicle/app/admin/backends/components/BackendForm.tsx
@@ -11,37 +11,9 @@ import { Input } from '@/components/ui/input'
 import { PasswordInput } from '@/components/ui/password-input'
 import { useEnvironment } from '@/app/context/environmentProvider'
 import { Textarea } from '@/components/ui/textarea'
+import { insertableBackendSchema } from '@/types/validation/backend'
 
-const formSchema = z.discriminatedUnion('providerType', [
-  z.object({
-    providerType: z.literal('openai'),
-    name: z.string().min(2, { message: 'Username must be at least 2 characters.' }),
-    apiKey: z.string().min(2, { message: 'Api Key  must be at least 2 characters.' }),
-  }),
-  z.object({
-    providerType: z.literal('anthropic'),
-    name: z.string().min(2, { message: 'Username must be at least 2 characters.' }),
-    apiKey: z.string().min(2, { message: 'Api Key  must be at least 2 characters.' }),
-  }),
-  z.object({
-    providerType: z.literal('logiclecloud'),
-    name: z.string().min(2, { message: 'Username must be at least 2 characters.' }),
-    endPoint: z.string().url(),
-    apiKey: z.string().min(2, { message: 'Api Key  must be at least 2 characters.' }),
-  }),
-  z.object({
-    providerType: z.literal('gcp-vertex'),
-    name: z.string().min(2, { message: 'Username must be at least 2 characters.' }),
-    credentials: z.string().min(2, { message: 'Credentials must be at least 2 characters.' }),
-  }),
-  z.object({
-    providerType: z.literal('perplexity'),
-    name: z.string().min(2, { message: 'Username must be at least 2 characters.' }),
-    apiKey: z.string().min(2, { message: 'Api Key  must be at least 2 characters.' }),
-  }),
-])
-
-export type BackendFormFields = z.infer<typeof formSchema>
+export type BackendFormFields = z.infer<typeof insertableBackendSchema>
 
 interface Props {
   backend: BackendFormFields
@@ -53,7 +25,7 @@ const BackendForm: FC<Props> = ({ backend, onSubmit, creating }) => {
   const { t } = useTranslation()
 
   const form = useForm<BackendFormFields>({
-    resolver: zodResolver(formSchema),
+    resolver: zodResolver(insertableBackendSchema),
     defaultValues: {
       ...backend,
     },
